wallet-core: Move some functionality from lib's root to `notes` module
- Add `NOTES_BUFFER_SIZE` to indicate the scratch buffer size for notes
- Add `MAX_INPUT_NOTES`, `phoenix_balance` and `BalanceInfo` under `notes` module
- Remove `MAX_INPUT_NOTES`, `phoenix_balance` and `BalanceInfo` from lib.rs
