+++ wallet-core/src/lib.rs
@@ -32,8 +32,8 @@ pub type Seed = [u8; 64];
 
 pub mod prelude {
     //! Re-export of the most commonly used types and traits.
+    pub use crate::input::MAX_INPUT_NOTES;
     pub use crate::keys;
-    pub use crate::{input::MAX_INPUT_NOTES, keys::RNG_SEED};
 }
 
 use alloc::vec::Vec;
@@ -44,62 +44,4 @@ use execution_core::transfer::phoenix::{Note, ViewKey as PhoenixViewKey};
 
 pub use notes::map_owned;
 
-/// Calculate the sum for all the given [`Note`]s that belong to the given
-/// [`PhoenixViewKey`].
-pub fn phoenix_balance<T>(
-    phoenix_vk: &PhoenixViewKey,
-    notes: impl Iterator<Item = T>,
-) -> BalanceInfo
-where
-    T: AsRef<Note>,
-{
-    let mut values: Vec<u64> = notes
-        .filter_map(|note| note.as_ref().value(Some(phoenix_vk)).ok())
-        .collect();
-
-    values.sort_by(|a, b| b.cmp(a));
-
-    let spendable = values.iter().take(input::MAX_INPUT_NOTES).sum();
-    let value =
-        spendable + values.iter().skip(input::MAX_INPUT_NOTES).sum::<u64>();
-
-    BalanceInfo { value, spendable }
-}
-
-/// Information about the balance of a particular key.
-#[derive(Debug, Default, Hash, Clone, Copy, PartialEq, Eq)]
-pub struct BalanceInfo {
-    /// The total value of the balance.
-    pub value: u64,
-    /// The maximum _spendable_ value in a single transaction. This is
-    /// different from `value` since there is a maximum number of notes one can
-    /// spend.
-    pub spendable: u64,
-}
-
-impl Serializable<{ 2 * u64::SIZE }> for BalanceInfo {
-    type Error = dusk_bytes::Error;
-
-    fn from_bytes(buf: &[u8; Self::SIZE]) -> Result<Self, Self::Error>
-    where
-        Self: Sized,
-    {
-        let mut reader = &buf[..];
-
-        let value = u64::from_reader(&mut reader)?;
-        let spendable = u64::from_reader(&mut reader)?;
-
-        Ok(Self { value, spendable })
-    }
-
-    #[allow(unused_must_use)]
-    fn to_bytes(&self) -> [u8; Self::SIZE] {
-        let mut buf = [0u8; Self::SIZE];
-        let mut writer = &mut buf[..];
-
-        writer.write(&self.value.to_bytes());
-        writer.write(&self.spendable.to_bytes());
-
-        buf
-    }
-}
+pub use notes::{map_owned, phoenix_balance, BalanceInfo};
